refactor: extract pkg/apis into independent Go module
    Make pkg/apis independently importable so external consumers can use
    kube-ovn API types (CRDs, status types, etc.) without pulling in the
    full kube-ovn dependency tree (OVS/OVN, CNI plugins, container runtimes,
    etc.).

    Key changes:

    - Add pkg/apis/go.mod declaring module github.com/kubeovn/kube-ovn/pkg/apis
      with minimal dependencies (k8s.io/api, k8s.io/apimachinery, k8s.io/klog)

    - Move BigInt type from pkg/internal/ to pkg/apis/types/ so it resides
      within the apis module boundary. pkg/internal/big_int.go now re-exports
      it as a type alias for full backward compatibility.

    - Inline OVN string constants in vpc.go (PolicyRouteAction values) and
      security-group.go (SgPolicy values) to eliminate the dependency on
      pkg/ovsdb/ovnnb. The values are unchanged: allow, drop,
      reroute, pass.

    - Update ippool.go and subnet.go to import BigInt from pkg/apis/types
      instead of pkg/internal.

    - Add replace directive in root go.mod to resolve the nested module
      locally during development.

    - Fix big_int_test.go to use keyed field literals (required because
      BigInt is now a type alias).
